SEQUOIADBMAINSTREAM-6075 Global backup flag
This is part of the restore to point-in-time feature.

Backups need to be identifiable as "global" to guarantee that the
backup operation was performed across the cluster. This is done by
adding a flag in the backup file header.

This change converts the backup operation type
engine::_barBackupHeader::_opType to a bitmap so it represents flags
and adds a new flag BAR_BACKUP_OP_TYPE_GLOBAL.

Note that this breaks the backup file compatibility with older versions
of the server and the sdbrestore tool because code was checking for
type equivalency. Older versions will fail to identify new backup files
as BAR_BACKUP_OP_TYPE_FULL vs BAR_BACKUP_OP_TYPE_INC.