SEQUOIADBMAINSTREAM-6367 Backup dir permission chk
Add a check that the backup directory is accessible for read/write.
This fixes a bug where a new directory can be created but changing
its permissions fails, then when the command is rerun the dir exists
and the create and permission check is skipped.

Cherry-pick: f493800ff58855ed092e61c2443bd2d41ea87ca7